Model and UserController restructuring
