--- backend/Models/SurveyCreationModel.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace backend.Models
-{
-    public class SurveyCreationModel
-    {
-        public int SurveyCreator { get; set; }
-        public string? SurveyName { get; set; }
-        public string? SurveyDescription { get; set; }
-    }
-}
