+++ backend/Models/UserModel.cs
@@ -5,12 +5,17 @@ using System.Threading.Tasks;
 
 namespace backend.Models
 {
-    public class UserModel
+    public class SurveyCreationModel
     {
-        public int? Id { get; set; }
-        public string? FullName { get; set; }
-        public string? Email { get; set; }
-        public string? Password { get; set; }
-        public string? Role { get; set; }
+        public int SurveyCreator { get; set; }
+        public string? SurveyName { get; set; }
+        public string? SurveyDescription { get; set; }
+    }
+
+    public class QuestionCreationModel
+    {
+        public int SurveyId { get; set; }
+        public string? QuestionText { get; set; }
+        public string? AnswerType { get; set; }
     }
 }
